flights/requirements.txt
functions-framework==3.*
SQLAlchemy
PyMySQL
pandas
pysqlite3-binary
requests